[artf4967] : Review TeamForge functionality for Solaris
The CTF integration hook scripts for Unix are not running in Solaris. This commit changes the implementation of AbstractCommandExcecutor.createHookScript() to be a bit more configurable by platform. The UnixCommandExectutor now creates two-line variable declaration and export which will work on Solaris and Linux.

* webapp/../AbstractCommandExecutor.java (createHookScript, getEnvironmentVariableString)
changed the implementation of createHookScript to depend on concrete subclass impl of new method, getEnvironmentVariableString

* webapp/../WindowsCommandExecutor.java
* webapp/../UnixCommandExecutor.java
(getEnvironmentVariableString, getEnvironmentSetCommand)
removed the getEnvironmentSetCommand and implemented new abstract method getEnviornmentVariableString() , which takes name, value arguments and returns the appropriate syntax for platform ( bat or shell script). This makes it easier to do two-command set and export for Solaris.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1544 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f