[artf5800] This revision patches the integration server's integration.war in a way that the output of any log statement used by its internal logger "Logger.java" goes to the Console. The output will be printed out to Subversion Edge's console and, if configured, will appear in the console.log.
The pattern for the messages follow the same as used on Grails, but uses the token [INTEGRATION] as an identifier of the thread/class that prints the output (an example is posted in the associated artifact).

ATTENTION: In order to get this done, I've added log4j's configuration into the integration.war manually, the pom.xml code is not correctly placing it into the war as expected.

* log4j.properties
- The configuration added manually into the war.

* Logger.java
- Updating the proxied method calls to use the System.out.println instead of log4j's.
- Adding the private method that prints the log message.

* BootstrapServlet.java
- Updating the integration server to display the appHome directory during bootstrap.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1993 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f