[artf6968] : In TeamForge mode against 5.4, viewvc doesn't work
After converting from SoapPy to Suds lib for the soap client in 2.2, viewvc no longer works when managed by TF 5.4. What happens is we attempt to connect to a 60 endpoint and then fallback to 50 if it fails, but Suds behaves slightly different than SoapPy, so an error was thrown before we expect it.

* integration-scripts/SourceForge.py (getDefaultSoapVersion)
moved SOAP client creation inside the try/except block so that the error can be caught.


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2762 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f